refac(api): migrate header component to TypeScript with improved structure and type safety
